Renames file for consistency
Renames `waitUntilSharedResourcesLoading.ts` to
`wait-until-shared-resources-loading.ts` to improve
naming consistency.
