+++ src/js/game/game-procedure/on-game-action/on-private-match-guest-start.ts
@@ -3,7 +3,7 @@ import { waitTime } from "../../../wait/wait-time";
 import { PrivateMatchGuestStart } from "../../game-actions/private-match-guest-start";
 import { GameProps } from "../../game-props";
 import { bindPlayerSelectAccordingToConfig } from "../bind-player-select-according-to-config";
-import { waitUntilSharedResourcesLoading } from "../waitUntilSharedResourcesLoading";
+import { waitUntilSharedResourcesLoading } from "../wait-until-shared-resources-loading";
 
 /** オプション */
 type Options = {
